test(ai): split phase_planner_economy_filter_test below-quota-peace groups into sibling file (Refs #2509)
Unblocks PR #2741 by lifting the two new EXPAND below-quota peace
rebuild-trap groups (resolvePhaseEconomyExpandBelowQuotaPeace*Active)
into a sibling test file so both files stay under the repo-lint
`repo.dart_file_non_comment_line_size` 1000-non-comment-line cap
(SPEC/program/repo-lint.md, SPEC/program/dart-file-non-comment-line-size.md).

- Move the two new groups (17 tests covering structural matrix, legacy
  parity, sibling-slot guards, three-call determinism, and the
  mutual-exclusion partition pin) into
  `phase_planner_economy_filter_below_quota_peace_test.dart`. Tests are
  byte-for-byte equivalent to the originals; no behavioral change.
- Duplicate the four file-private `_colonialAcquisitionPopulated`,
  `_colonialMilitaryPopulated`, `_colonialNavalPopulated`, and
  `_colonialCivilianPopulated` const fixtures used by the moved groups
  (keeps the sibling file self-contained without exposing fixture
  symbols through a shared library).
- Drop the now-unused `colonial_pressure.dart`,
  `kBelowQuotaPeaceMinRegimentsBeforeDeclareWar`, and
  `kObserverConquestMinOwProvincesPerGp` imports from the original
  file; add a header pointer to the sibling so readers can find the
  EXPAND below-quota peace resolver tests.

Verification
- `dart run tool/ct_repo_lint.dart --rule repo.dart_file_non_comment_line_size`: pass.
- `dart analyze test/planning/phase_planner_economy_filter_test.dart test/planning/phase_planner_economy_filter_below_quota_peace_test.dart`: no issues.
- `dart test test/planning/phase_planner_economy_filter_test.dart test/planning/phase_planner_economy_filter_below_quota_peace_test.dart`: 47/47 pass (30 original + 17 lifted).
- `dart test packages/colonizethis_ai`: 1163/1163 pass, 2 skipped (no regressions).

Made with [Cursor](https://cursor.com)
